refactor: move 시 put 메서드 대신 patch 메서드 사용

+++ src/test/java/chess/service/ChessServiceTest.java
@@ -36,7 +36,7 @@ public class ChessServiceTest {
     @BeforeEach
     void setUp() {
         chessDao = new ChessDaoImpl(jdbcTemplate);
-        gameId = chessDao.initGame("칙촉조시제이", "123").intValue();
+        gameId = chessDao.initGame("칙촉조시제이", "123");
         chessService = new ChessService(chessDao);
     }
